backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.1
lxml==6.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-docx==0.8.11
sqlparse==0.5.3
typing_extensions==4.14.1
